+++ packages/Pagination/src/Pagination.styles.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const Pagination = styled.nav`
+  align-items: center;
+  display: flex;
+  text-align: center;
+`;