+++ packages/ListBox/CHANGELOG.md
@@ -1,8 +1,10 @@
-# CHANGELOG
+# Changelog
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
+
+## [Unreleased]
 
 ## [0.9.0] - 2020-06-16
 
@@ -10,3 +12,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add onKeyUp event for `SPACE`, `ESC` and `ENTER` before onKeyDown was handling all key interaction now its only responsible for `↑`,`↓` keys.
   this might break testing that depends onKeyDown event.
+
+## [0.9.3] - 2020-08-13
+
+### Changed
+
+- Add `label` prop on `ListBox.Trigger` component to support custom text as the label.
+
+## [0.9.4] - 2020-08-17
+
+### Fixed
+
+- Update `lastKnownSelectedOptions` in the state, when `useListBox.types.updateOptions` has been dispatched.